+++ mail/mail-uniffi/CHANGELOG.md
@@ -10,6 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
   - [ET-736] Human Verification
+  - [ET-2438] PGP/Mime attachment forward + Reply support.
+
+### Fixed
+
+  - Draft body not being saved
 
 
 ## [0.63.0] - 2025-03-24